refactor(lib): move classification helpers beside their traits
The Elixir keyword-Call, Tcl command-name and Python alias helpers were
defined in metrics::cognitive / metrics::npa yet imported by the Checker
and Getter impls, so the parse layer depended on the metric layer. They
now live in a lang_helpers module the metrics import instead.

SpaceKind and HalsteadType are Getter return types, so they move out of
spaces / metrics::halstead into their own modules; both stay re-exported
at their public paths. The two termcolor helpers leave tools.rs for
output::color, and FromPathError leaves error.rs so that file holds only
the parse-layer MetricsError.

Preparatory step for #1376: no public path changes, no metric moves.
