+++ src/metrics/npm.rs
@@ -20,9 +20,10 @@
 use std::fmt;
 
 use crate::checker::{Checker, csharp_accessor_count};
+use crate::lang_helpers::python::python_is_block;
 use crate::langs::*;
 use crate::macros::implement_metric_trait;
-use crate::metrics::npa::{accessibility_ratio, python_is_block, ts_member_is_public};
+use crate::metrics::npa::{accessibility_ratio, ts_member_is_public};
 use crate::node::Node;
 use crate::*;
