+++ src/output/color.rs
@@ -16,7 +16,7 @@
 
 use std::io::{StdoutLock, Write};
 
-use termcolor::{Buffer, BufferWriter, ColorChoice, ColorSpec, WriteColor};
+use termcolor::{Buffer, BufferWriter, Color, ColorChoice, ColorSpec, WriteColor};
 
 /// Whether the terminal dump serializers ([`crate::dump_root`],
 /// [`crate::dump_ops`], [`crate::dump_node`],
@@ -241,6 +241,21 @@ where
     let rendered = render(&mut chunked);
     chunked.emit_pending()?;
     rendered
+}
+
+// Accept `&mut dyn WriteColor` rather than `&mut StandardStreamLock` so
+// tests (e.g. `function::dump_spans`) can substitute `termcolor::NoColor`
+// over a `Vec<u8>` to capture the rendered bytes. Production callers
+// continue to pass `&mut StandardStreamLock`, which unsized-coerces to
+// the trait object at the call site.
+#[inline]
+pub(crate) fn color(stdout: &mut dyn WriteColor, color: Color) -> std::io::Result<()> {
+    stdout.set_color(ColorSpec::new().set_fg(Some(color)))
+}
+
+#[inline]
+pub(crate) fn intense_color(stdout: &mut dyn WriteColor, color: Color) -> std::io::Result<()> {
+    stdout.set_color(ColorSpec::new().set_fg(Some(color)).set_intense(true))
 }
 
 #[cfg(test)]
